ssh-prompt > Added a template for how to set defined users — demonstrated
### Rationale
If this template is used and placed in the current folder that
ssh-connect is called from it will pull in the variables to help define
what available users are for the system being connected to.  Its a
usability feature for when you develop on different systems and don't
want to remember every defined user for each system.