vnc-prompt > Created VNC prompt scripts that utilize a tunneled ssh connection — demonstrated
### Rationale
In order to create secure VNC connections, it is recommended to pipe the
VNC connection through a built SSH tunnel.  This script automates that
process so you don't have to look up how to do this every time.  Instead
the prompts guide you through and build the command for you.  As an added
benefit you don't have to be running the vncserver all the time, but
with this script it will start up a vncserver for you first. Since it
creates a dynamic configuration file on the fly, It incorporates a
control master socket connection to only connect to the end box twice,
and all other connections are hidden in the tunnel.