use-ssh-agents > Added separate SSH-agent for script — incomplete
### Rationale
Future proofing again so that the user will not have to re-authenticate
their key multiple times during the use of the script since in the future
there will be potentially multiple separate ssh connections made before
a stable long term one is created.

Note this does not address ssh-agent death.  It should die with the
parent script.