remove-server-autostart > Relocated & disabled vncserver startup — demonstrated
### Rationale
This prevents the annoying autostart of another server when there might
be one that already exists on the remote host.  This is also future
proofing as this changes around the order in which the port forward is
applied.  If enabled there would be 2 different ssh connections made
unless the control master still is alive.