fix/vnc-cli-order > Changed workflow order
\### Rationale
The logical order of input questions does not line up with how the user
should expect it to.  For VNC connections, now the ssh connection is
determined first to include the ssh-add key then the VNC settings are
defined. The SSH connection will ask for the ip and then the port
before determining the username to login with.